Refactored map part into its own component
