+++ practice_app/components/MapView.module.css
@@ -1,0 +1,5 @@
+.mapContainer {
+    position: relative;
+    width: 100vw;
+    height: 300px;
+}
